srcs/backend/requirements.txt
Django>=4.2.1,<4.3.0
psycopg2-binary
python-decouple
django-environ
djangorestframework
django-cors-headers
Pillow
djangorestframework-simplejwt>=4.7.2